pyproject.toml
[tool.poetry]
name = "gemini"
version = "1.1.0"
description = "Raci Backend App"
authors = ["Antony <antony.muchiri@cema.africa>"]
readme = "README.md"
homepage = "https://github.com/yourusername/gemini"
repository = "https://github.com/yourusername/gemini"
documentation = "https://github.com/yourusername/gemini/README.md"
license = "MIT"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"                   # change to ^3.12 or ^3.13 if you prefer
Django = "^5.1"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0"
django-extensions = "^3.2"
djangorestframework = { version = "^3.15", optional = true }
drf-spectacular = { version = "^0.27", optional = true }
celery = { version = "^5.4", optional = true }
redis = { version = "^5.0", optional = true }
gunicorn = "^23.0"
whitenoise = "^6.7"
djangorestframework-simplejwt = "^5.5.1"
django-filter = "^25.2"
django-cors-headers = "^4.9.0"
django-auditlog = "^3.3.0"
pillow = "^12.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8"
isort = "^5.13"
flake8 = "^7.1"
mypy = "^1.11"
pytest = "^8.3"
pytest-django = "^4.8"
pytest-cov = "^5.0"
pre-commit = "^3.8"
django-debug-toolbar = "^4.4"
ipython = "^8.27"
python-semantic-release = "^10.5.2"

[tool.poetry.extras]
api = ["djangorestframework", "drf-spectacular"]
tasks = ["celery", "redis"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# ────────────────────────── Tool configurations ──────────────────────────

[tool.black]
line-length = 99
target-version = ['py311']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 99
known_first_party = ["gemini", "config", "core"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_equality = true
plugins = ["mypy_django_plugin.main"]

[tool.mypy.overrides]
module = "config.*"
django_settings_module = "config.settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = "tests.py test_*.py *_tests.py"
addopts = "-ra -q --cov=."

[tool.coverage.run]
source = ["."]
omit = ["*/migrations/*", "*/tests/*", "manage.py"]


# ────────────────────────── Semantic Release ──────────────────────────

[tool.semantic_release]
version_toml = ["pyproject.toml:tool.poetry.version"]
branch = "main"
upload_to_vcs_release = true
build_command = ""
major_on_zero = false
tag_format = "v{version}"

commit_parser = "conventional"

[tool.semantic_release.commit_parser_options]
allowed_tags = [
    "build", "chore", "ci", "docs", "feat",
    "fix", "perf", "style", "refactor", "test"
]
minor_tags = ["feat"]
patch_tags = ["fix", "perf", "docs"]  

[tool.semantic_release.branches.main]
match = "(main|master)"
prerelease = false

[tool.semantic_release.changelog]
template_dir = "templates"
changelog_file = "CHANGELOG.md"
exclude_commit_patterns = ["chore*", "ci*"]

[tool.semantic_release.changelog.environment]
block_start_string = "{%"
block_end_string = "%}"
variable_start_string = "{{"
variable_end_string = "}}"
comment_start_string = "{#"
comment_end_string = "#}"
newline_sequence = "\n"
